refactor config.rs
