+++ frb_codegen/src/config/mod.rs
@@ -1,0 +1,4 @@
+pub(crate) mod raw_opts;
+pub(crate) mod opts;
+pub mod opts_parser;
+pub(crate) mod raw_opts_parser;
