+++ frb_codegen/src/config/opts.rs
@@ -1,0 +1,137 @@
+use anyhow::{Context, Result};
+use std::fs;
+use std::path::{Path, PathBuf};
+use crate::ir::IrFile;
+use crate::parser;
+use crate::utils::misc::BlockIndex;
+
+#[derive(Debug, Clone, serde::Serialize)]
+pub struct Opts {
+    pub rust_input_path: String,
+    pub dart_output_path: String,
+    pub dart_decl_output_path: Option<String>,
+    pub c_output_path: Vec<String>,
+    pub rust_crate_dir: String,
+    pub rust_output_path: String,
+    pub class_name: String,
+    pub dart_format_line_length: u32,
+    pub dart_enums_style: bool,
+    pub skip_add_mod_to_lib: bool,
+    pub llvm_path: Vec<String>,
+    pub llvm_compiler_opts: String,
+    pub manifest_path: String,
+    pub dart_root: Option<String>,
+    pub build_runner: bool,
+    pub block_index: BlockIndex,
+    pub skip_deps_check: bool,
+    pub wasm_enabled: bool,
+    pub inline_rust: bool,
+}
+
+impl Opts {
+    pub fn get_ir_file(&self) -> Result<IrFile> {
+        // info!("Phase: Parse source code to AST");
+        let source_rust_content = fs::read_to_string(&self.rust_input_path).with_context(|| {
+            format!(
+                "Failed to read rust input file \"{}\"",
+                self.rust_input_path
+            )
+        })?;
+        let file_ast = syn::parse_file(&source_rust_content).unwrap();
+
+        // info!("Phase: Parse AST to IR");
+
+        Ok(parser::parse(
+            &source_rust_content,
+            file_ast,
+            &self.manifest_path,
+        ))
+    }
+
+    pub fn dart_api_class_name(&self) -> String {
+        self.class_name.clone()
+    }
+
+    pub fn dart_api_impl_class_name(&self) -> String {
+        format!("{}Impl", self.class_name)
+    }
+
+    pub fn dart_wire_class_name(&self) -> String {
+        format!("{}Wire", self.class_name)
+    }
+
+    pub fn dart_platform_class_name(&self) -> String {
+        format!("{}Platform", self.class_name)
+    }
+
+    pub fn dart_wasm_module(&self) -> String {
+        format!("{}WasmModule", self.class_name)
+    }
+
+    pub(crate) fn dart_output_root(&self) -> Option<&str> {
+        Path::new(
+            self.dart_decl_output_path
+                .as_ref()
+                .unwrap_or(&self.dart_output_path),
+        )
+            .file_stem()?
+            .to_str()
+    }
+
+    pub fn dart_wasm_output_path(&self) -> PathBuf {
+        Path::new(&self.dart_output_path).with_extension("web.dart")
+    }
+
+    pub fn dart_io_output_path(&self) -> PathBuf {
+        Path::new(&self.dart_output_path).with_extension("io.dart")
+    }
+
+    pub fn dart_common_output_path(&self) -> PathBuf {
+        Path::new(&self.dart_output_path).with_extension("common.dart")
+    }
+
+    pub fn rust_wasm_output_path(&self) -> PathBuf {
+        Path::new(&self.rust_output_path).with_extension("web.rs")
+    }
+
+    pub fn rust_io_output_path(&self) -> PathBuf {
+        Path::new(&self.rust_output_path).with_extension("io.rs")
+    }
+
+    pub fn dart_root_or_default(&self) -> String {
+        self.dart_root
+            .clone()
+            .unwrap_or_else(|| env!("CARGO_MANIFEST_DIR").to_string())
+    }
+
+    pub fn dart_freezed_path(&self) -> PathBuf {
+        PathBuf::from(
+            self.dart_decl_output_path
+                .as_deref()
+                .unwrap_or(&self.dart_output_path),
+        )
+            .with_extension("freezed.dart")
+    }
+
+    pub fn get_rust_output_paths(&self) -> PathForGeneration {
+        PathForGeneration {
+            base_path: PathBuf::from(self.rust_output_path.clone()),
+            io_path: self.rust_io_output_path(),
+            wasm_path: self.rust_wasm_output_path(),
+        }
+    }
+
+    pub fn get_dart_output_paths(&self) -> PathForGeneration {
+        PathForGeneration {
+            base_path: PathBuf::from(self.dart_output_path.clone()),
+            io_path: self.dart_io_output_path(),
+            wasm_path: self.dart_wasm_output_path(),
+        }
+    }
+}
+
+pub struct PathForGeneration {
+    pub base_path: PathBuf,
+    pub io_path: PathBuf,
+    pub wasm_path: PathBuf,
+}
